fixed vercel error moving authOptions to a separate file, updated imports and removed remaining console.log calls
