+++ src/app/(main)/profile/[id]/page.tsx
@@ -1,7 +1,7 @@
 import { getServerSession } from "next-auth";
 import UpdateUserForm from "./form";
 import { redirect } from "next/navigation";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/lib/authOptions";
 
 //metadata.title = "Profile Update";
